requirements.txt
loguru==0.7.3
python-dotenv==1.0.1
pydantic==2.10.6
pdfplumber==0.11.5
langchain-ollama==0.2.3
